beam cylinder moved to gnomon
