+++ nexus_constructor/instrument_view.py
@@ -58,39 +58,20 @@ class InstrumentView(QWidget):
         self.cube_entity = Qt3DCore.QEntity(self.component_root_entity)
         self.cube_mesh = Qt3DExtras.QCuboidMesh()
 
-        self.num_neutrons = 9
-
-        # Create a dictionary for neutron-related objects so that they are always in scope and not destroyed by C++
-        self.neutron_objects = {
-            "entities": [],
-            "meshes": [],
-            "transforms": [],
-            "animation_controllers": [],
-            "animations": [],
-        }
-
-        for _ in range(self.num_neutrons):
-            self.neutron_objects["entities"].append(
-                Qt3DCore.QEntity(self.gnomon_root_entity)
-            )
-            self.neutron_objects["meshes"].append(Qt3DExtras.QSphereMesh())
-            self.neutron_objects["transforms"].append(Qt3DCore.QTransform())
-
         self.beam_cylinder_length = 40
 
         # Create the gnomon resources and get its camera
         self.gnomon = Gnomon(
-            self.gnomon_root_entity, self.view.camera(), self.add_qcomponents_to_entity
+            self.gnomon_root_entity,
+            self.view.camera(),
+            self.beam_material,
+            self.grey_material,
+            self.add_qcomponents_to_entity,
         )
         self.gnomon_camera = self.gnomon.get_gnomon_camera()
 
         self.create_layers()
         self.initialise_view()
-
-        # Initialise beam objects
-        self.cylinder_entity = Qt3DCore.QEntity(self.component_root_entity)
-        self.cylinder_mesh = Qt3DExtras.QCylinderMesh()
-        self.cylinder_transform = Qt3DCore.QTransform()
 
         # Dictionaries for component-related objects also to prevent them from going out of scope
         self.component_meshes = {}
@@ -98,7 +79,7 @@ class InstrumentView(QWidget):
         self.component_transformations = {}
 
         # Insert the beam cylinder last. This ensures that the semi-transparency works correctly.
-        self.setup_beam_cylinder()
+        self.gnomon.setup_beam_cylinder()
 
         # Move the gnomon when the camera view changes
         self.view.camera().viewVectorChanged.connect(self.gnomon.update_gnomon)
@@ -262,130 +243,6 @@ class InstrumentView(QWidget):
         for component in components:
             entity.addComponent(component)
 
-    @staticmethod
-    def set_cylinder_mesh_dimensions(cylinder_mesh, radius, length, rings):
-        """
-        Sets the dimensions of a cylinder mesh.
-        :param cylinder_mesh: The cylinder mesh to modify.
-        :param radius: The desired radius.
-        :param length: The desired length.
-        :param rings: The desired number of rings.
-        """
-        cylinder_mesh.setRadius(radius)
-        cylinder_mesh.setLength(length)
-        cylinder_mesh.setRings(rings)
-
-    @staticmethod
-    def set_beam_transform(cylinder_transform):
-        """
-        Configures the transform for the beam cylinder by giving it a matrix. The matrix will turn the cylinder sideways
-        and then move it "backwards" in the z-direction by 20 units so that it ends at the location of the sample.
-        :param cylinder_transform: A QTransform object.
-        """
-        cylinder_matrix = QMatrix4x4()
-        cylinder_matrix.rotate(270, QVector3D(1, 0, 0))
-        cylinder_matrix.translate(QVector3D(0, 20, 0))
-
-        cylinder_transform.setMatrix(cylinder_matrix)
-
-    def setup_beam_cylinder(self):
-        """
-        Sets up the beam cylinder by giving the cylinder entity a mesh, a material, and a transformation.
-        """
-        self.set_cylinder_mesh_dimensions(
-            self.cylinder_mesh, 2.5, self.beam_cylinder_length, 2
-        )
-        self.set_beam_transform(self.cylinder_transform)
-        self.add_qcomponents_to_entity(
-            self.cylinder_entity,
-            [self.cylinder_mesh, self.beam_material, self.cylinder_transform],
-        )
-
-    @staticmethod
-    def set_sphere_mesh_radius(sphere_mesh, radius):
-        """
-        Sets the radius of a sphere mesh.
-        :param sphere_mesh: The sphere mesh to modify.
-        :param radius: The desired radius.
-        """
-        sphere_mesh.setRadius(radius)
-
-    @staticmethod
-    def set_neutron_animation_properties(
-        neutron_animation,
-        neutron_animation_controller,
-        animation_distance,
-        time_span_offset,
-    ):
-        """
-        Prepares a QPropertyAnimation for a neutron by giving it a target, a distance, and loop settings.
-        :param neutron_animation: The QPropertyAnimation to be configured.
-        :param neutron_animation_controller: The related animation controller object.
-        :param animation_distance: The starting distance of the neutron.
-        :param time_span_offset: The offset that allows the neutron to move at a different time from other neutrons.
-        """
-        neutron_animation.setTargetObject(neutron_animation_controller)
-        neutron_animation.setPropertyName(b"distance")
-        neutron_animation.setStartValue(0)
-        neutron_animation.setEndValue(animation_distance)
-        neutron_animation.setDuration(500 + time_span_offset)
-        neutron_animation.setLoopCount(-1)
-        neutron_animation.start()
-
-    def setup_neutrons(self):
-        """
-        Sets up the neutrons and their animations by preparing their meshes and then giving offset and
-        distance parameters to an animation controller.
-        """
-
-        # Create lists of x, y, and time offsets for the neutron animations
-        x_offsets = [0, 0, 0, 2, -2, 1.4, 1.4, -1.4, -1.4]
-        y_offsets = [0, 2, -2, 0, 0, 1.4, -1.4, 1.4, -1.4]
-        time_span_offsets = [0, -5, -7, 5, 7, 19, -19, 23, -23]
-
-        neutron_radius = 1.5
-
-        for i in range(self.num_neutrons):
-
-            self.set_sphere_mesh_radius(
-                self.neutron_objects["meshes"][i], neutron_radius
-            )
-
-            neutron_animation_controller = NeutronAnimationController(
-                x_offsets[i] * 0.5,
-                y_offsets[i] * 0.5,
-                self.neutron_objects["transforms"][i],
-            )
-            neutron_animation_controller.set_target(
-                self.neutron_objects["transforms"][i]
-            )
-
-            x_cylinder_length = 8
-
-            neutron_animation = QPropertyAnimation(
-                self.neutron_objects["transforms"][i]
-            )
-            self.set_neutron_animation_properties(
-                neutron_animation,
-                neutron_animation_controller,
-                x_cylinder_length,
-                time_span_offsets[i],
-            )
-
-            self.neutron_objects["animation_controllers"].append(
-                neutron_animation_controller
-            )
-            self.neutron_objects["animations"].append(neutron_animation)
-
-            self.add_qcomponents_to_entity(
-                self.neutron_objects["entities"][i],
-                [
-                    self.neutron_objects["meshes"][i],
-                    self.grey_material,
-                    self.neutron_objects["transforms"][i],
-                ],
-            )
-
     def initialise_view(self):
         """
         Calls the methods for defining materials, setting up the sample cube, and setting up the neutrons. Beam-related
@@ -393,5 +250,5 @@ class InstrumentView(QWidget):
         """
         self.give_colours_to_materials()
         self.setup_sample_cube()
-        self.setup_neutrons()
         self.gnomon.create_gnomon()
+        self.gnomon.setup_neutrons()
